+++ packages/api/src/extrinsic/extensions/NoopSignedExtension.ts
@@ -1,3 +1,4 @@
+import { PortableRegistry } from '@dedot/codecs';
 import { SignerPayloadJSON } from '@dedot/types';
 import { SignedExtension } from './SignedExtension.js';
 import { ISubstrateClient } from '../../types.js';
@@ -31,4 +32,32 @@ export class NoopSignedExtension extends SignedExtension {
   toPayload(): Partial<SignerPayloadJSON> {
     return {}; // No payload contribution
   }
+}
+
+/**
+ * Checks if a type is an empty struct or tuple (doesn't require external input).
+ * 
+ * @param registry The portable registry
+ * @param typeId The type ID to check
+ * @returns True if the type is an empty struct or tuple, false otherwise
+ */
+export function isEmptyStructOrTuple(registry: PortableRegistry, typeId: number): boolean {
+  try {
+    const type = registry.findType(typeId);
+    
+    // Check if it's an empty struct
+    if (type.typeDef.type === 'Struct' && type.typeDef.value.fields.length === 0) {
+      return true;
+    }
+    
+    // Check if it's an empty tuple
+    if (type.typeDef.type === 'Tuple' && type.typeDef.value.fields.length === 0) {
+      return true;
+    }
+  } catch (error) {
+    // Ignore errors
+  }
+  
+  // All other cases (including errors) require input
+  return false;
 }
